move getLocale to model and import to smoothly-display also
