+++ src/model/index.ts
@@ -1,6 +1,7 @@
 export { Autocomplete } from "./Autocomplete"
 export { Color } from "./Color"
 export { Fill } from "./Fill"
+export { getLocale } from "./getLocale"
 export { Icon } from "./Icon"
 export { Message } from "./Message"
 export { Notice } from "./Notice"
